change float to double

+++ HandWrite/PPoint.h
@@ -2,42 +2,42 @@
 class Point
 {
 public:
-   float m_x;
-   float m_y;
+   double m_x;
+   double m_y;
 
    Point() :m_x(0), m_y(0) {};
-   Point(float x, float y) :m_x(x), m_y(y){};
+   Point(double x, double y) :m_x(x), m_y(y){};
 
 
-   float x() const { return m_x; }
-   float y() const { return m_y; }
+   double x() const { return m_x; }
+   double y() const { return m_y; }
 
-   float& rx() { return m_x; }
-   float& ry() { return m_y; }
+   double& rx() { return m_x; }
+   double& ry() { return m_y; }
 
    //! Compare two points at subpixel resolution
    static bool roughlySame(const Point& p1, const Point &p2) {
-      float dx = p1.x() - p2.x();
-      float dy = p1.y() - p2.y();
-      float d = dx*dx + dy*dy;
+      double dx = p1.x() - p2.x();
+      double dy = p1.y() - p2.y();
+      double d = dx*dx + dy*dy;
       return d <= ((1 / 4.0)*(1 / 4.0));
    }
    bool roughlySame(const Point& point) const { return roughlySame(*this, point); }
 
    //! Are the two points less than one pixel different?
    static bool intSame(const Point &p1, const Point &p2) {
-      float dx = p1.x() - p2.x();
-      float dy = p1.y() - p2.y();
-      float d = dx*dx + dy*dy;
+      double dx = p1.x() - p2.x();
+      double dy = p1.y() - p2.y();
+      double d = dx*dx + dy*dy;
       return d < 1.0;
    }
    bool intSame(const Point &point) const { return intSame(*this, point); }
 
-   static float distance(const Point &p1, const Point &p2)
+   static double distance(const Point &p1, const Point &p2)
    {
       return hypot(p1.x() - p2.x(), p1.y() - p2.y());
    }
-   float distance(const Point &point) const { return distance(*this, point); }
+   double distance(const Point &point) const { return distance(*this, point); }
 
 };
 
@@ -61,48 +61,48 @@ static inline bool operator==(const Point& p1, const Point& p2)
 class PPoint
 {
 public:
-   float m_x;
-   float m_y;
-   float m_p; // Ñ¹Á¦
+   double m_x;
+   double m_y;
+   double m_p; // Ñ¹Á¦
 
    PPoint() :m_x(0), m_y(0), m_p(0) {};
-   PPoint(float x, float y, float p = 1.0) :m_x(x), m_y(y), m_p(p) {};
-   PPoint(const Point& pt, float p = 1.0) :m_x(pt.x()), m_y(pt.y()), m_p(p) {};
+   PPoint(double x, double y, double p = 1.0) :m_x(x), m_y(y), m_p(p) {};
+   PPoint(const Point& pt, double p = 1.0) :m_x(pt.x()), m_y(pt.y()), m_p(p) {};
    ~PPoint() {};
 
-   float x() const { return m_x; }
-   float y() const { return m_y; }
-   float pressure() const { return m_p; }
+   double x() const { return m_x; }
+   double y() const { return m_y; }
+   double pressure() const { return m_p; }
 
-   float& rx() { return m_x; }
-   float& ry() { return m_y; }
+   double& rx() { return m_x; }
+   double& ry() { return m_y; }
 
    //! Set this point's pressure value
-   void setPressure(float p) { assert(p >= 0 && p <= 1); m_p = p; }
+   void setPressure(double p) { assert(p >= 0 && p <= 1); m_p = p; }
 
    //! Compare two points at subpixel resolution
    static bool roughlySame(const PPoint& p1, const PPoint &p2) {
-      float dx = p1.x() - p2.x();
-      float dy = p1.y() - p2.y();
-      float d = dx*dx + dy*dy;
+      double dx = p1.x() - p2.x();
+      double dy = p1.y() - p2.y();
+      double d = dx*dx + dy*dy;
       return d <= ((1 / 4.0)*(1 / 4.0));
    }
    bool roughlySame(const PPoint& point) const { return roughlySame(*this, point); }
 
    //! Are the two points less than one pixel different?
    static bool intSame(const PPoint &p1, const PPoint &p2) {
-      float dx = p1.x() - p2.x();
-      float dy = p1.y() - p2.y();
-      float d = dx*dx + dy*dy;
+      double dx = p1.x() - p2.x();
+      double dy = p1.y() - p2.y();
+      double d = dx*dx + dy*dy;
       return d < 1.0;
    }
    bool intSame(const PPoint &point) const { return intSame(*this, point); }
 
-   static float distance(const PPoint &p1, const PPoint &p2)
+   static double distance(const PPoint &p1, const PPoint &p2)
    {
       return hypot(p1.x() - p2.x(), p1.y() - p2.y());
    }
-   float distance(const PPoint &point) const { return distance(*this, point); }
+   double distance(const PPoint &point) const { return distance(*this, point); }
 
 };
